requirements.txt
pandas==2.2.3
jupyter==1.0.0
ipykernel==6.29.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
connectorx==0.3.3
matplotlib==3.9.2
pyarrow==17.0.0
seaborn==0.13.2
fastapi[standard]==0.115.2
pydantic==2.9.2
streamlit==1.39.0
uvicorn==0.32.0
jinja2
python-multipart
torch
nltk
gensim
faiss-cpu
